# Improve performance on GPU