I only updated `BICGSTAB` and  `CGS` for the moment.
In the future, we should remove all occurrence of broadcast in the other solvers.

```julia
using Random, LinearAlgebra, CUDA, Krylov, BenchmarkTools
CUDA.allowscalar(false)
Random.seed!(0)
n = 2^12
A = Matrix(Symmetric(rand(n, n)))
cA = CuArray(A)
b = ones(n)
cb = CuArray(b)
@benchmark bicgstab($cA, $cb)
```

Before this pull request:
```julia
BenchmarkTools.Trial: 
  memory estimate:  121.50 MiB
  allocs estimate:  2654329
  --------------
  minimum time:     6.909 s (0.18% GC)
  median time:      6.909 s (0.18% GC)
  mean time:        6.909 s (0.18% GC)
  maximum time:     6.909 s (0.18% GC)
  --------------
  samples:          1
  evals/sample:     1
```

Now: 
```julia
BenchmarkTools.Trial: 
  memory estimate:  6.25 MiB
  allocs estimate:  368761
  --------------
  minimum time:     6.460 s (0.00% GC)
  median time:      6.460 s (0.00% GC)
  mean time:        6.460 s (0.00% GC)
  maximum time:     6.460 s (0.00% GC)
  --------------
  samples:          1
  evals/sample:     1
 ```