@dpo 2020-11-28
ps: maybe open an issue so we remember to make similar changes to other solvers?